requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
xgboost>=2.0.0
tensorflow>=2.15.0
optuna>=3.5.0
pyyaml>=6.0
openpyxl>=3.1.0
pyarrow>=14.0.0
